Moved the image component into its own sub-directory.
